Switch to using cpp_new_array instead of plain malloc.
This was supposed to fix the weird off-by-one error in which the goto-trace specifies that the array { x = 0, x = 10, x = 20 } violates the assertion (!(a[0].x == 10 && a[1].x == 20 && a[1].x == 30)).

It didn't; the error persists, but I'll keep this version of the code since it gets to use java_array_new instead of malloc in an untried form; hopefully this is less likely to cause trouble in future.