Implement mocking of classes with an opaque base
This is achieved by intercepting the construction of the derived class, since actually replacing a superconstructor is considerably more difficult. While the whole object is mocked, the fields can nontheless only be assigned values consistent with the execution of the opaque superconstructor followed by the known actions of the derived constructor -- so if the opaque base initialises x and y in an unknown way, then the dervied class overwrites y with 0, CBMC should never propose a solution that assigns y any other value (but remains free to pick x).