fix for no java entry point
The refactoring of `get_main_symbol` returned a tuple type, but in the
original version, it could also return directly which was not possible
after the refacoring. Therefore not specifying an entry point in a class
without main could result in a violated assertion.

We now return a 3-tuple with an additional Boolean that signals whether
to return after `get_main_symbol` or to continue.