emit warning if too many answers are demanded
The answer list in mocking only holds a limited number of results. If
the test case execution deviates from the CBMC trace, we might get more
calls to the mocked object. This patch reverts the index counter and
emits are warning. This is temporary until a better solution / more
precise analysis is found.