Fix invalid maybe null initialization of Strings
The initialization of String arrays was broken.
It was impossible to create a String array that
is non-null or maybe-null.
When jbmc is run without a valid model for the
String class (which makes little sense anyways)
then a String object cannot be instantiated in
the harness; the initialization will be null in
this case.